Created BP_PathBlockerManager
It is currently in existance under assumption it will be used to tell each PathBlocker weather they are meant to be active or not. Tho the use much be limited as the moment as PathBlockers can decide themselves weather then need to be active or not, the plan is that in the future multi-level pathBlockers could need the use. For now this remains empty